requirements.txt
annotated-types==0.7.0
anyio==4.10.0
av==15.0.0
cachetools==5.5.2
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
colorama==0.4.6
coloredlogs==15.0.1
ctranslate2==4.6.0
distro==1.9.0
faster-whisper==1.2.0
filelock==3.19.1
flatbuffers==25.2.10
fsspec==2025.7.0
google-auth==2.40.3
google-auth-oauthlib==1.2.2
groq==0.31.0
gspread==6.2.1
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.34.4
humanfriendly==10.0
idna==3.10
mpmath==1.3.0
numpy==2.3.2
oauth2client==4.1.3
oauthlib==3.3.1
onnxruntime==1.22.1
packaging==25.0
protobuf==6.32.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAudio==0.2.14
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pyparsing==3.2.3
pyreadline3==3.5.4
python-dotenv==1.1.1
PyYAML==6.0.2
requests==2.32.4
requests-oauthlib==2.0.0
rsa==4.9.1
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.2
SpeechRecognition==3.14.3
sympy==1.14.0
tokenizers==0.21.4
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0





#to install requirements run the following command:-   pip install -r requirements.txt